+++ packages/credential-sdk/src/vc/crypto/common/DockCryptoSignature.js
@@ -8,7 +8,7 @@ import jsonld from 'jsonld';
 import jsigs from 'jsonld-signatures';
 
 import stringify from 'json-stringify-deterministic';
-import semver from 'semver/preload';
+import semver from 'semver/preload.js';
 import { u8aToU8a } from '../../../utils/bytes';
 import { withExtendedStaticProperties } from '../../../utils/inheritance';
 import CustomLinkedDataSignature from './CustomLinkedDataSignature';