+++ packages/dock-blockchain-modules/src/attest/index.js
@@ -1,0 +1,3 @@
+export * from './actions';
+export { default as DockAttestModuleInternal } from './internal';
+export { default as DockAttestModule } from './module';