+++ packages/dock-blockchain-modules/src/blob/module.js
@@ -1,13 +1,12 @@
-import { NoBlobError } from '@docknetwork/credential-sdk/modules/abstract/blob/errors';
+import { NoBlobError, AbstractBlobModule } from '@docknetwork/credential-sdk/modules/abstract/blob';
 import { option } from '@docknetwork/credential-sdk/types/generic';
 import { DockBlobId } from '@docknetwork/credential-sdk/types';
-import { AbstractBlobModule } from '@docknetwork/credential-sdk/modules/abstract/blob';
 import { injectDock } from '../common';
 import { OwnerWithBlob } from './types';
 import DockBlobModuleInternal from './internal';
 
 /** Class to create and update Blobs on chain. */
-export default class BlobModule extends injectDock(AbstractBlobModule) {
+export default class DockBlobModule extends injectDock(AbstractBlobModule) {
   static DockOnly = DockBlobModuleInternal;
 
   /**