+++ packages/dock-blockchain-modules/src/common/dock-api-provider.js
@@ -1,6 +1,6 @@
 import { ApiProvider } from '@docknetwork/credential-sdk/modules/abstract/common';
-import { ensureInstanceOf } from '@docknetwork/credential-sdk/utils/type-helpers';
-import { DockDIDModuleInternal } from '../did/internal';
+import { ensureInstanceOf } from '@docknetwork/credential-sdk/utils';
+import DockDIDModuleInternal from '../did/internal';
 
 class DockApiProvider extends ApiProvider {
   constructor(dock) {