+++ packages/dock-blockchain-api/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @docknetwork/dock-blockchain-api
 
+## 0.4.0
+
+### Minor Changes
+
+- Fix exports
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.5.0
+
 ## 0.3.0
 
 ### Minor Changes